Move SuperAdminSearchContents component to SuperAdmin common component folder
